RF: Move CameraDevice to main module again, remove old file
